+++ src/main/java/com/najasin/domain/manual/question/entity/Question.java
@@ -1,14 +1,12 @@
-package com.najasin.domain.myManual.question.entity;
+package com.najasin.domain.manual.question.entity;
 
-import com.najasin.domain.comment.entity.Comment;
+import com.najasin.domain.manual.dto.response.JffMyQuestion;
 import com.najasin.domain.user.entity.userType.UserType;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
 
 @Getter
 @Entity
@@ -27,11 +25,11 @@ public class Question {
     @Enumerated(EnumType.STRING)
     private QuestionType questionType;
 
-    @OneToMany(mappedBy = "question")
-    private List<Comment> comments;
-
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_type_id", referencedColumnName = "user_type_id")
     private UserType userType;
 
+    public JffMyQuestion toJffMyQuestion() {
+        return new JffMyQuestion(id, question);
+    }
 }
