--- src/main/java/com/najasin/domain/myManual/question/repository/QuestionRepository.java
@@ -1,12 +1,0 @@
-package com.najasin.domain.myManual.question.repository;
-
-import com.najasin.domain.myManual.question.entity.Question;
-import com.najasin.domain.myManual.question.entity.QuestionType;
-import com.najasin.domain.user.entity.userType.UserType;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface QuestionRepository extends JpaRepository<Question, Long> {
-    List<Question> getQuestionsByQuestionTypeAndUserType(QuestionType questionType, UserType userType);
-}
