+++ src/main/java/com/najasin/domain/userKeyword/entity/UserKeyword.java
@@ -1,6 +1,6 @@
 package com.najasin.domain.userKeyword.entity;
 
-import com.najasin.domain.myManual.keyword.entity.Keyword;
+import com.najasin.domain.manual.keyword.entity.Keyword;
 import com.najasin.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
